@@ -1,5 +1,5 @@
 module github.com/gorilla/handlers
 
-go 1.14
+go 1.20
 
-require github.com/felixge/httpsnoop v1.0.1
+require github.com/felixge/httpsnoop v1.0.3
