@@ -21,12 +21,11 @@ type canonical struct {
 //
 // Example:
 //
-//  r := mux.NewRouter()
-//  canonical := handlers.CanonicalHost("http://www.gorillatoolkit.org", 302)
-//  r.HandleFunc("/route", YourHandler)
+//	r := mux.NewRouter()
+//	canonical := handlers.CanonicalHost("http://www.gorillatoolkit.org", 302)
+//	r.HandleFunc("/route", YourHandler)
 //
-//  log.Fatal(http.ListenAndServe(":7000", canonical(r)))
-//
+//	log.Fatal(http.ListenAndServe(":7000", canonical(r)))
 func CanonicalHost(domain string, code int) func(h http.Handler) http.Handler {
 	fn := func(h http.Handler) http.Handler {
 		return canonical{h, domain, code}
